Send frames to Link via FrameListener interface
I was having trouble figuring out the interaction between Link's `send`
and FrameListener's `onFrame`, but then I realized that they're actually
the same thing. Note that NetworkAdapter's `send` is different from
Link's method of the same name, It's for the user of NetworkAdapter.

The user of NetworkAdapter sends a tuple of (address, data) but receives
Frames (by passing a FrameListener into the NetworkAdapter). Can this
asymmetry be resolved?